Adds test for handling Mount
